+++ src/core/service-management/service-management.service.ts
@@ -4,12 +4,17 @@ import { OAuthService } from '../oauth/oauth.service';
 import { ActivityAPI } from './activity-api.service';
 import { CompositeBulkAPI } from './composite-bulk-api.service';
 import { CompositeTreeAPI } from './composite-tree-api.service';
+import { ServiceCallAPI } from './service-call-api.service';
 
-
-
+/**
+ * Main service for accessing SAP Field Service Management Service Management APIs.
+ * Provides access to activity, service call, and composite operations.
+ */
 export class ServiceManagementAPIService {
 
     public activity: ActivityAPI;
+
+    public serviceCall: ServiceCallAPI;
 
     public composite: {
         tree: CompositeTreeAPI,
@@ -22,6 +27,7 @@ export class ServiceManagementAPIService {
         private _auth: Readonly<OAuthService>
     ) {
         this.activity = new ActivityAPI(this._config, this._http, this._auth)
+        this.serviceCall = new ServiceCallAPI(this._config, this._http, this._auth)
         this.composite = {
             tree: new CompositeTreeAPI(this._config, this._http, this._auth),
             bulk: new CompositeBulkAPI(this._config, this._http, this._auth),